--- src/ode/methods/mod.rs
@@ -1,8 +1,0 @@
-//! NumericalMethods for Ordinary Differential Equations (ODEs).
-
-// Adams Methods for solving ordinary differential equations.
-pub mod adams;
-
-// Initial step size determination methods for ODEs.
-mod h_init;
-pub use h_init::h_init;
